nexus/packages/edge/pyproject.toml
[project]
name = "nexus-edge-server"
version = "0.1.0"
description = "NEXUS Local Edge Model Server"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn>=0.30.0",
    "structlog>=24.1.0",
    "pydantic>=2.7.0",
]

[project.optional-dependencies]
gpu = [
    "llama-cpp-python>=0.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
